requirements.txt
aiogram==3.24.0
aiohttp==3.13.3
gTTS==2.5.4
pydub==0.25.1
python-dotenv==1.2.1

